+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
+## 2.7.12
+
+## Enhancements
+- Add customer info to paywall info and tracked events
+- Add stripe/paddle intro offer eligibility
+
+## Fixes
+- Fix dismiss animation for bottom sheets and modals on newer Samsung devices
 
 ## 2.7.11